% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_helpers.R
\name{create_allcomb_df}
\alias{create_allcomb_df}
\title{Create combination of all queries and all references}
\usage{
create_allcomb_df(query_names, reference_names)
}
\arguments{
\item{query_names}{Character vector of query names}

\item{reference_names}{Character vector of reference names}
}
\value{
A data frame with columns 'query' and 'reference' where each row is a unique pairing of a query and a reference
}
\description{
Create combination of all queries and all references
}
