% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_helpers.R
\name{create_qbestd_df}
\alias{create_qbestd_df}
\title{Create results data frame}
\usage{
create_qbestd_df(search_mf, search_results)
}
\arguments{
\item{search_mf}{Search manifest with pairs of query and references searched}

\item{search_results}{Matrix of scores, start and end indices returned by DTW search}
}
\value{
A data frame with columns 'query', 'reference', and 'score'
}
\description{
Create results data frame
}
