% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{dist_stdeuc}
\alias{dist_stdeuc}
\title{Calculate standardized Euclidean distances between query and reference feature matrices}
\usage{
dist_stdeuc(query_feats, ref_feats)
}
\arguments{
\item{query_feats}{Feature matrix for query, of shape M rows and F columns}

\item{ref_feats}{Feature matrix for reference, of shape N rows and F columns}
}
\value{
A distance matrix with M rows and N columns
}
\description{
Calculate standardized Euclidean distances between query and reference feature matrices
}
