% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{fast_scale}
\alias{fast_scale}
\title{A fast scale function for calcuating the standardised Euclidean}
\usage{
fast_scale(feature_matrix)
}
\description{
Adapted from \href{https://www.r-bloggers.com/2016/02/a-faster-scale-function/}{https://www.r-bloggers.com/2016/02/a-faster-scale-function/}
}
