% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npz_helpers.R
\name{fetch_npz_item}
\alias{fetch_npz_item}
\title{Fetch array by item name from NumPy array archive}
\usage{
fetch_npz_item(npz_file, item_name)
}
\arguments{
\item{npz_file}{Path to .npz file}

\item{item_name}{Name within archive dictionary}
}
\value{
A feature matrix with T rows and F columns, where
T is the number of time steps and F the number of
feature components
}
\description{
Fetch array by item name from NumPy array archive
}
