% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npz_helpers.R
\name{fetch_npz_names}
\alias{fetch_npz_names}
\title{Fetch filenames from named NumPy array archive}
\usage{
fetch_npz_names(npz_file)
}
\arguments{
\item{npz_file}{Path to .npz file}
}
\value{
A character vector of names within the NumPy array archive
}
\description{
Fetch filenames from named NumPy array archive
}
