% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_helpers.R
\name{norm_rf2014}
\alias{norm_rf2014}
\title{Normalize distance matrix according to procedure proposed by Rodriguez-Fuentes et al. (2014)}
\usage{
norm_rf2014(qr_dists)
}
\arguments{
\item{qr_dists}{A distance matrix}
}
\value{
A normalized distance matrix
}
\description{
For formal description of procedure see: \href{https://doi.org/10.1109/ICASSP.2014.6855122}{https://doi.org/10.1109/ICASSP.2014.6855122}.
}
