% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbd_std.R
\name{qbe_std}
\alias{qbe_std}
\title{Detect queries in test items}
\usage{
qbe_std(
  queries_loc,
  references_loc,
  names_fetcher = fetch_npz_names,
  features_fetcher = fetch_npz_item,
  search_mf_maker = create_allcomb_df,
  post_processor = create_qbestd_df,
  nndtw_func = sousrir_1nndtw,
  ssdtw_func = sousrir_ssdtw,
  progress_bar = TRUE,
  use_multisession = TRUE
)
}
\arguments{
\item{queries_loc}{Location of queries (default: an npz file containing a named dictionary of NumPy feature matrices of shape TxF)}

\item{references_loc}{Location of references (default: an npz file containing a named dictionary of NumPy feature matrices of shape TxF)}

\item{names_fetcher}{A function that takes queries_loc/references_loc and returns the items contained in them (default: \link{fetch_npz_names})}

\item{features_fetcher}{A function that takes queries_loc/references_loc and an item name, and returns the features associated with that item (default: \link{fetch_npz_item})}

\item{search_mf_maker}{A function that takes the list of query and reference names and returns a two-column data frame with pairs of queries and references (default: \link{create_allcomb_df})}

\item{post_processor}{A function to process the search results (default: \link{create_qbestd_df})}

\item{nndtw_func}{A function to shortlist starting indices (default: \link{sousrir_1nndtw})}

\item{ssdtw_func}{A function to calculate a score of how likely a query occurs in a reference, given starting indices (default: \link{sousrir_ssdtw})}

\item{progress_bar}{Show progress bar while running search}

\item{use_multisession}{Use future::multisession to run search using multiple R sessions in parallel}
}
\description{
Detect queries in test items
}
