% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp2sec_helpers.R
\name{samp2sec_libmfcc}
\alias{samp2sec_libmfcc}
\title{Convert Librosa MFCC sample number into seconds}
\usage{
samp2sec_libmfcc(sample_number, sample_rate, hop_length = 512)
}
\arguments{
\item{sample_number}{An integer vector of sample numbers to convert}

\item{sample_rate}{Sample rate of original audio}

\item{hop_length}{Hop length used when extracting MFCC features (default: 512)}
}
\value{
A numeric vector of time in seconds corresponding to the sample numbers
}
\description{
Convert Librosa MFCC sample number into seconds
}
