% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw_helpers.R
\name{sousrir_1nndtw}
\alias{sousrir_1nndtw}
\title{Return index of 1-nearest neighbour search using rundtw function from IncDTW package}
\usage{
sousrir_1nndtw(query_feats, ref_feats)
}
\arguments{
\item{query_feats}{Feature matrix for query}

\item{ref_feats}{Feature matrix for reference}
}
\value{
An integer indicating the start of the best match for the query within the reference
}
\description{
Return index of 1-nearest neighbour search using rundtw function from IncDTW package
}
