% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw_helpers.R
\name{sousrir_ssdtw}
\alias{sousrir_ssdtw}
\title{Return score of how likely a query occurs in a reference, given a starting index}
\usage{
sousrir_ssdtw(
  query_name,
  ref_name,
  query_feats,
  ref_feats,
  top_match_start,
  min_match_ratio = 0.5,
  max_match_ratio = 2,
  distance_func = dist_scipy_stdeuc,
  distnorm_func = norm_rf2014,
  return_dtwalign = FALSE
)
}
\arguments{
\item{query_name}{Name of query (for error reporting in case function fails)}

\item{ref_name}{Name of reference (for error reporting in case function fails)}

\item{query_feats}{Query feature matrix (of shape M rows and F columns)}

\item{ref_feats}{Reference feature matrix (of shape N rows and F columns)}

\item{top_match_start}{Top match location returned by \link{sousrir_1nndtw}}

\item{min_match_ratio}{Minimum match length as ratio of query (default: 0.5 = half the query size)}

\item{max_match_ratio}{Maximum match length as ratio of query (default: 2.0 = twice the query size)}

\item{distance_func}{Function to compute distances between query and reference (default: \link{dist_stdeuc})}

\item{distnorm_func}{Function to normalize computed distances (default: \link{norm_rf2014})}

\item{return_dtwalign}{Whether or not to return alignment object (i.e. for plotting alignment)}
}
\value{
A numeric score of how likely the query occurs in the reference (or NA if no acceptable alignment was found given the parameters)
}
\description{
Return score of how likely a query occurs in a reference, given a starting index
}
